Rename project and hone goals
I think "kindling" reflects the intention of the project. I considered
`react-firestarter` which is kinda cutesy with the interplay of
firebase, but I may move away from firebase eventually.

It doesn't look like I can change the Firebase project ID. I can
recreate it later with the right name.

Using the Firebase logo for the moment until I can get a logo for the
project.